Further renaming to reduce ambiguity
