+++ src/main/java/frc/robot/limelights/VisionSubsystem.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.vision;
+package frc.robot.limelights;
 
 import java.util.Arrays;
 import java.util.HashSet;
@@ -49,7 +49,7 @@ public class VisionSubsystem extends SubsystemBase {
     }
 
     /** Latest Limelight data. May contain faulty data unsuitable for odometry. */
-    private LimelightData[] limelightDatas = new LimelightData[2];
+    private VisionData[] limelightDatas = new VisionData[2];
     /** Last heartbeat of the front LL (updated every frame) */
     private long lastHeartbeatFrontLL = 0;
     /** Last heartbeat of the back LL (updated every frame) */
@@ -91,11 +91,11 @@ public class VisionSubsystem extends SubsystemBase {
     @Override
     public void periodic() {
         // This method gets data in about 4 to 8 ms.
-        LimelightData[] filteredLimelightDatas = getFilteredLimelightData(false);
+        VisionData[] filteredLimelightDatas = getFilteredLimelightData(false);
 
         // This loop generally updates data in about 6 ms, but may double or triple for no apparent reason.
         // This causes loop overrun warnings, however, it doesn't seem to be due to inefficient code and thus can be ignored.
-        for (LimelightData data : filteredLimelightDatas) {
+        for (VisionData data : filteredLimelightDatas) {
             if (data.canTrustRotation) {
                 // Only trust rotational data when adding this pose.
                 TunerConstants.DriveTrain.setVisionMeasurementStdDevs(VecBuilder.fill(9999999, 9999999, 1));
@@ -126,7 +126,7 @@ public class VisionSubsystem extends SubsystemBase {
      * @apiNote Will theoretically stop updating data if the heartbeat resets.
      * However, this happens at 2e9 frames, which would take consecutive 96 days at a consistent 240 fps.
      */
-    private LimelightData[] getFilteredLimelightData(boolean useStored) {
+    private VisionData[] getFilteredLimelightData(boolean useStored) {
         LimelightHelpers.PoseEstimate frontLLDataMT2 = null;
         LimelightHelpers.PoseEstimate backLLDataMT2 = null;
         long heartbeatFrontLL = -1;
@@ -148,14 +148,14 @@ public class VisionSubsystem extends SubsystemBase {
             if (heartbeatFrontLL == -1 || this.lastHeartbeatFrontLL < heartbeatFrontLL) {
                 frontLLDataMT2 = LimelightHelpers.getBotPoseEstimate_wpiBlue_MegaTag2(LimelightConstants.FRONT_APRIL_TAG_LL);
                 LimelightHelpers.PoseEstimate frontLLDataMT = LimelightHelpers.getBotPoseEstimate_wpiBlue(LimelightConstants.FRONT_APRIL_TAG_LL);
-                this.limelightDatas[0] = new LimelightData(LimelightConstants.FRONT_APRIL_TAG_LL, frontLLDataMT, frontLLDataMT2);
+                this.limelightDatas[0] = new VisionData(LimelightConstants.FRONT_APRIL_TAG_LL, frontLLDataMT, frontLLDataMT2);
                 this.lastHeartbeatFrontLL = heartbeatFrontLL == -1 ? this.lastHeartbeatFrontLL : heartbeatFrontLL;
             }
             
             if (heartbeatBackLL == -1 || this.lastHeartbeatBackLL < heartbeatBackLL) {
                 backLLDataMT2 = LimelightHelpers.getBotPoseEstimate_wpiBlue_MegaTag2(LimelightConstants.BACK_APRIL_TAG_LL);
                 LimelightHelpers.PoseEstimate backLLDataMT = LimelightHelpers.getBotPoseEstimate_wpiBlue(LimelightConstants.BACK_APRIL_TAG_LL);
-                this.limelightDatas[1] = new LimelightData(LimelightConstants.BACK_APRIL_TAG_LL, backLLDataMT, backLLDataMT2);
+                this.limelightDatas[1] = new VisionData(LimelightConstants.BACK_APRIL_TAG_LL, backLLDataMT, backLLDataMT2);
                 this.lastHeartbeatBackLL = heartbeatBackLL == -1 ? this.lastHeartbeatBackLL : heartbeatBackLL;
             }
 
@@ -169,7 +169,7 @@ public class VisionSubsystem extends SubsystemBase {
                 || ((frontLLDataMT2 != null && frontLLDataMT2.tagCount == 0) && (backLLDataMT2 != null && backLLDataMT2.tagCount == 0))
                 || (frontLLDataMT2 != null && frontLLDataMT2.avgTagDist > VisionConstants.TRUST_TAG_DISTANCE
                     && backLLDataMT2 != null && backLLDataMT2.avgTagDist > VisionConstants.TRUST_TAG_DISTANCE)) {
-                return new LimelightData[0];
+                return new VisionData[0];
             }
         }
 
@@ -192,7 +192,7 @@ public class VisionSubsystem extends SubsystemBase {
             }
         }
         if (frontLLDataMT2 == null && backLLDataMT2 == null) {
-            return new LimelightData[0];
+            return new VisionData[0];
         }
 
         // Returns the data with the greater tag count.
@@ -202,24 +202,24 @@ public class VisionSubsystem extends SubsystemBase {
             && (backLLDataMT2 == null
                 || backLLDataMT2.avgTagDist > VisionConstants.TRUST_TAG_DISTANCE
                 || frontLLDataMT2.tagCount > backLLDataMT2.tagCount)) {
-                return new LimelightData[]{ this.limelightDatas[0] };
+                return new VisionData[]{ this.limelightDatas[0] };
         }
         else if ((backLLDataMT2 != null && (useStored || this.lastHeartbeatBackLL == heartbeatBackLL))
             && (frontLLDataMT2 == null
                 || frontLLDataMT2.avgTagDist > VisionConstants.TRUST_TAG_DISTANCE
                 || backLLDataMT2.tagCount > frontLLDataMT2.tagCount)) {
-                return new LimelightData[]{ this.limelightDatas[1] };
+                return new VisionData[]{ this.limelightDatas[1] };
         }
 
         // Returns the data that's closer to its respective camera than 90% of the other's distance.
         // 90% is a heuristic.
         if ((!useStored && this.lastHeartbeatFrontLL == heartbeatFrontLL)
             && frontLLDataMT2.avgTagDist < backLLDataMT2.avgTagDist * 0.9) {
-            return new LimelightData[]{ this.limelightDatas[0] };
+            return new VisionData[]{ this.limelightDatas[0] };
         }
         else if ((!useStored && this.lastHeartbeatBackLL == heartbeatBackLL)
             && backLLDataMT2.avgTagDist < frontLLDataMT2.avgTagDist * 0.9) {
-            return new LimelightData[]{ this.limelightDatas[1] };
+            return new VisionData[]{ this.limelightDatas[1] };
         }
 
         // This return statement assumes that both LLs have the same amount of tags and
@@ -232,7 +232,7 @@ public class VisionSubsystem extends SubsystemBase {
      */
     private void optimizeLimelights() {
         byte index = 0; // Used only for setting the optimized flag, so that this can be a for-each loop.
-        for (LimelightData limelightData : this.limelightDatas) {
+        for (VisionData limelightData : this.limelightDatas) {
             if (limelightData == null || limelightData.optimized) {
                 return;
             }
@@ -376,7 +376,7 @@ public class VisionSubsystem extends SubsystemBase {
     }
     
     /**
-     * This is a helper for {@link VisionSubsystem#optimizeLimelights(LimelightData[])} smart cropping.
+     * This is a helper for {@link VisionSubsystem#optimizeLimelights(VisionData[])} smart cropping.
      * @param id - The target ID to consider.
      * @return Whether to expect another tag on the left, right, or neither.
      * @apiNote Left : -1 ; Right : +1 ; Neither : 0.
@@ -404,7 +404,7 @@ public class VisionSubsystem extends SubsystemBase {
      * @apiNote If MegaTag rotation cannot be trusted, it will use the odometry's current rotation.
      */
     public Pose2d getEstimatedPose() {
-        LimelightData[] filteredLimelightDatas = getFilteredLimelightData(true);
+        VisionData[] filteredLimelightDatas = getFilteredLimelightData(true);
 
         if (filteredLimelightDatas.length == 0) {
             System.err.println("getEstimatedPose() | NO LIMELIGHT DATA, DEFAULTING TO EMTPY POSE2D");
